update Compatibility matrix
The Compatibility matrix didn't mention RTRlib (https://github.com/rtrlib/rtrlib/), which is a C library implementing the RTR protocol. Part of the software package is the rtrclient, which dumps valid ROA payloads, similar to RTRdump.

RPKI support in FRRouting and BIRD is based on RTRlib. Currently, SSH transport in the RTRlib only allows for key-based authentication. As such, FRRouting does not support password-based SSH but key-based SSH.