feat: update cf.pub key and cache file path
The private key that encrypts the file in `https://rpki.cloudflare.com/rpki.json` is being rotated.
In order to avoid any downtime, we created a second file with the new encryption key in `https://rpki.cloudflare.com/v2/rpki.json`.
In this PR, we update the path for the cache file, so we use the newly encrypted v2/rpki.json, and also the new public key in cf.pub that matches it.

The old file will also need to be updated so we can deprecate the old encryption keys.
You need to download this new release to make sure your code does not break when the key is updated. Alternatively, you can continue to use the release you are using now and simply update cf.pub and pass the -cache flag with the correct url.

DEADLINE: 18-03-2024 !!!!